requirements.txt
hyper
googletrans
duolingo-api
tqdm
spacy
selenium
gtts
playsound
colorlog
PySide2
unidecode